+++ src/GraphQL.StarWars.TypeFirst/Types/StarWarsQuery.cs
@@ -1,6 +1,4 @@
-using GraphQL.StarWars.TypeFirst.Types;
-
-namespace GraphQL.StarWars.TypeFirst;
+namespace GraphQL.StarWars.TypeFirst.Types;
 
 [Name("Query")]
 public class StarWarsQuery
